Make "3D coords generating" message less scary
Raising an exception (which was done previously) prints a
traceback in the notebook, and it makes it look as if a
serious error occurred.

We do not want to give people that impression, since the only
thing that has happened is that 3D coordinate generation started.
So print a simple message instead.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>