Do not pull images when running tasks with docker
This commit results in tasks running without pulling images.

It currently works for docker, but it probably does not work
for singularity or shifter.

It skips the generation of the image_uri, so, rather than using
the image_uri, we need to create the run commands manually. This
still needs to be done for singularity and shifter.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>