Add path and extension for singularity sif files
This attempts to add a path and the ".sif" extension for the
singularity sif files. It expects the ".sif" file to already be
residing in the singularity directory before the taskflow is ran.

This is untested.

We will need to add another taskflow soon that will pull images, create
singularity files, etc. when the user asks it to be done.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>