Added basic openchemistry-client in python
The client uses girder-client to make requests. Currently, it can only
list molecules.

There is a sub-command structure in place. After "openchemistry-client",
"molecules" is currently a command dealing with molecules (one other
planned command is "calculations").

After the "molecules" subcommand, the user can do things like list
molecules with an "ls" command. The full command for this is
"openchemistry-client molecules ls". It prints out a formatted
list of molecules.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>