Added ability to upload molecules from a file
Molecules can be uploaded from a file like the following:
```
openchemistry-client --api-key $MY_API_KEY molecules upload ethanol.smi
```

The user can specify a format, but if a format is not specified,
it tries to use the file extension.

Also, if the GIRDER_API_KEY environment variable is set, that will
be used by default for the api key, so it would not need to be
passed in via the CLI in that case.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>