Added ability to submit and download calculations
The submission of calculations re-uses functions and classes from
openchemistry. The girder client that openchemistry is using is
patched to be the girder client that the openchemistry client is
using.

An example of an input json file was placed in
`examples/calculations/input.json`. A molecule ID is required.

Downloading a calculation saves the cjson to an output file.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>