Add a way to get optimized geometry IDs
With a CalculationResult object (normally, they are named `result` in
the notebook examples), you can now call the property
`.optimized_geometry_id` on it to attempt to get the optimized geometry
ID. If the optimized geometry ID is found, it will be returned.

Only optimization tasks will have an optimized geometry ID. All others
will print 'None' and return None.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>