Warn when no task is provided to run_calculations
Unlike the API for performing optimizations on single molecules,
such as `mol.optimize()` and `mol.energy()`, `oc.run_calculations()`
is a general function which does not automatically insert the task into
the input_parameters dict.

If there is no task provided, it was likely an accident. Warn the user
when no task was provided. No task will typically mean an "energy"
calculation will be performed.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>