Added a taskflow to pull containers
This taskflow will be used by an image manager in mongochemclient.
The user will be able to select the type of container, the
repository, and the tag.

It seems to work for docker just fine. Hopefully it will work for
shifter.

Singularity is not completely working. Although the image appears to
be built, I am not actually sure where it is being placed. There
may be an error at the end of building for an invalid directory
or something.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>